+++ src/components/data.js
@@ -1,8 +1,9 @@
 import moment from "moment-timezone"
 import Africa from "./timezones/Africa"
 import Europe from "./timezones/Europe"
+import Russia from "./timezones/Russia"
 
-const timezonesRaw = [...Africa, ...Europe]
+const timezonesRaw = [...Africa, ...Europe, ...Russia]
 
 const Now = moment().utc().format("x")
 const rawZones = timezonesRaw.map(tz => {
