+++ src/components/timezones/EastAsia.js
@@ -19,10 +19,6 @@ export const China = {
       code: "MN"
     },
     {
-      title: "Russia: Buryatia, Irkutsk Oblast",
-      code: "RU"
-    },
-    {
       title: "Taiwan",
       code: "TW"
     }
@@ -42,12 +38,60 @@ export const JapanKorea = {
       code: "KP"
     },
     {
-      title: "Russia: Sakha (western part), Zabaykalsky Krai, Amur Oblast",
-      code: "RU"
-    },
-    {
       title: "South Korea",
       code: "KR"
+    }
+  ]
+}
+
+// Krasnoyarsk Time - UTC+7
+export const Krasnoyarsk = {
+  city: "Krasnoyarsk",
+  zone: "Asia/Krasnoyarsk",
+  flags: [
+    {
+      title: "Mongolia: Bayan-Ölgii, Govi-Altai, Khovd, Uvs, Zavkhan",
+      code: "MN"
+    },
+  ]
+}
+
+// Omsk Time - UTC+6
+export const Omsk = {
+  city: "Omsk",
+  zone: "Asia/Omsk",
+  flags: [
+    {
+      title: "Kazakhstan: Akmola Region, Almaty Region, East Kazakhstan Region, Jambyl Region, Karaganda Region, Kostanay Region, Kyzylorda Region, North Kazakhstan Region, Pavlodar Region, Turkistan Region",
+      code: "KZ"
+    },
+    {
+      title: "Kyrgyzstan",
+      code: "KG"
+    },
+  ]
+}
+
+// Yekaterinburg Time - UTC+5
+export const Yekaterinburg = {
+  city: "Yekaterinburg",
+  zone: "Asia/Yekaterinburg",
+  flags: [
+    {
+      title: "Kazakhstan: Aktobe Region, Atyrau Region, Mangystau Region, West Kazakhstan Region",
+      code: "KZ"
+    },
+    {
+      title: "Tajikistan",
+      code: "TJ"
+    },
+    {
+      title: "Turkmenistan",
+      code: "TM"
+    },
+    {
+      title: "Uzbekistan",
+      code: "UZ"
     }
   ]
 }
