+++ src/components/timezones/Russia.js
@@ -1,155 +1,68 @@
-// Kaliningrad Time - UTC + 2
-export const Kaliningrad = {
-  city: "Kaliningrad",
-  zone: "Europe/Kaliningrad",
-  flags: [
-    {
-      title: "Russia: Oblast Kaliningrad",
-      code: "RU"
-    },
-  ]
-}
-
-// Kamchatka Time - UTC+12
-export const Kamchatka = {
-  city: "Kamchatka",
-  zone: "Asia/Kamchatka",
-  flags: [
-    {
-      title: "Russia: Kamchatka, Chukotka Autonomous Okrug",
-      code: "RU"
-    }
-  ]
-}
-
-// Krasnoyarsk Time - UTC+7
-export const Krasnoyarsk = {
-  city: "Krasnoyarsk",
-  zone: "Asia/Krasnoyarsk",
-  flags: [
-    {
-      title: "Mongolia: Bayan-Ölgii, Govi-Altai, Khovd, Uvs, Zavkhan",
-      code: "MN"
-    },
-    {
-      title: "Russia: Khakassia, Tuva Republic, Krasnoyarsk Region, Kemerovo Oblast, Novosibirsk Oblast, Tomsk Oblast",
-      code: "RU"
-    }
-  ]
-}
-
-// Moscow Time - UTC+3
-export const Moscow = {
-  city: "Moscow",
-  zone: "Europe/Moscow",
-  flags: [
-    {
-      title: "Belarus",
-      code: "BY"
-    },
-    {
-      title: "Russia: Moscow, Saint Petersburg, Sevastopol, Adygea, Chechnya, Chuvashia, Dagestan, Ingushetia, Kabardino-Balkaria, Kalmykia, Karachay-Cherkessia, Karelia, Komi, Mari El, Mordovia, North Ossetia-Alania, Tatarstan, Krasnodar Region, Stavropol Region, Arkhangelsk Oblast, Astrakhan Oblast, Belgorod Oblast, Bryansk Oblast, Ivanovo Oblast, Kaluga Oblast, Kirov Oblast, Kostroma Oblast, Kursk Oblast, Leningrad Oblast, Lipetsk Oblast, Moscow Oblast, Murmansk Oblast, Nizhnyi Novgorod Oblast, Novgorod Oblast, Oryol Oblast, Penza Oblast, Pskov Oblast, Ryazan Oblast, Saratov Oblast, Smolensk Oblast, Tambov Oblast, Tula Oblast, Tver Oblast, Vladimir Oblast, Volgograd Oblast, Vologda Oblast, Voronezh Oblast, Yaroslavl Oblast, Nenets Autonomous Okrug, Crimea",
-      code: "RU"
-    },
-    {
-      title: "Turkey",
-      code: "TR"
-    }
-  ]
-}
-
-// Omsk Time - UTC+6
-export const Omsk = {
-  city: "Omsk",
-  zone: "Asia/Omsk",
-  flags: [
-    {
-      title: "Kazakhstan: Akmola Region, Almaty Region, East Kazakhstan Region, Jambyl Region, Karaganda Region, Kostanay Region, Kyzylorda Region, North Kazakhstan Region, Pavlodar Region, Turkistan Region",
-      code: "KZ"
-    },
-    {
-      title: "Kyrgyzstan",
-      code: "KG"
-    },
-    {
-      title: "Russia: Altai Republic, Altai Region, Omsk Oblast",
-      code: "RU"
-    },
-  ]
-}
-
-// Samara Time - UTC+4
-export const Samara = {
-  city: "Samara",
-  zone: "Europe/Samara",
-  flags: [
-    {
-      title: "Armenia",
-      code: "AM"
-    },
-    {
-      title: "Azerbaijan",
-      code: "AZ"
-    },
-    {
-      title: "Georgia",
-      code: "GE"
-    },
-    {
-      title: "Russia: Udmurtia, Samara Oblast, Ulyanovsk Oblast",
-      code: "RU"
-    },
-  ]
-}
-
-// Srednekolymsk or Magadan Time - UTC+11
-export const Srednekolymsk = {
-  city: "Srednekolymsk",
-  zone: "Asia/Srednekolymsk",
-  flags: [
-    {
-      title: "Russia: Sakha (eastern part), Magadan Oblast, Sakhalin Oblast",
-      code: "RU"
-    }
-  ]
-}
-
-// Vladivostok Time - UTC+10
-export const Vladivostok = {
-  city: "Vladivostok",
-  zone: "Asia/Vladivostok",
-  flags: [
-    {
-      title: "Russia: Sakha (central part), Khabarovsk Region, Primorsky Krai, Jewish Autonomous Oblast",
-      code: "RU"
-    }
-  ]
-}
-
-// Yekaterinburg Time - UTC+5
-export const Yekaterinburg = {
-  city: "Yekaterinburg",
-  zone: "Asia/Yekaterinburg",
-  flags: [
-    {
-      title: "Kazakhstan: Aktobe Region, Atyrau Region, Mangystau Region, West Kazakhstan Region",
-      code: "KZ"
-    },
-    {
-      title: "Russia: Bashkortostan, Perm Region, Chelyabinsk Oblast, Kurgan Oblast, Orenburg Oblast, Sverdlovsk Oblast, Tyumen Oblast, Khanty-Mansi Autonomous Okrug, Yamalo-Nenets Autonomous Okrug",
-      code: "RU"
-    },
-    {
-      title: "Tajikistan",
-      code: "TJ"
-    },
-    {
-      title: "Turkmenistan",
-      code: "TM"
-    },
-    {
-      title: "Uzbekistan",
-      code: "UZ"
-    }
-  ]
-}
+export default [
+  {
+    country: "Russia",
+    subdiv: ["Kaliningrad Oblast"],
+    zone: "Europe/Kaliningrad",
+    flag: "RU"
+  },
+  {
+    country: "Russia",
+    subdiv: ["Moscow", "Saint Petersburg", "Sevastopol", "Adygea", "Chechnya", "Chuvashia", "Dagestan", "Ingushetia", "Kabardino-Balkaria", "Kalmykia", "Karachay-Cherkessia", "Karelia", "Komi", "Mari El", "Mordovia", "North Ossetia-Alania", "Tatarstan", "Krasnodar Region", "Stavropol Region", "Arkhangelsk Oblast", "Astrakhan Oblast", "Belgorod Oblast", "Bryansk Oblast", "Ivanovo Oblast", "Kaluga Oblast", "Kirov Oblast", "Kostroma Oblast", "Kursk Oblast", "Leningrad Oblast", "Lipetsk Oblast", "Moscow Oblast", "Murmansk Oblast", "Nizhnyi Novgorod Oblast", "Novgorod Oblast", "Oryol Oblast", "Penza Oblast", "Pskov Oblast", "Ryazan Oblast", "Saratov Oblast", "Smolensk Oblast", "Tambov Oblast", "Tula Oblast", "Tver Oblast", "Vladimir Oblast", "Volgograd Oblast", "Vologda Oblast", "Voronezh Oblast", "Yaroslavl Oblast", "Nenets Autonomous Okrug", "Crimea"],
+    zone: "Europe/Moscow",
+    flag: "RU"
+  },
+  {
+    country: "Russia",
+    subdiv: ["Udmurtia", "Samara Oblast", "Ulyanovsk Oblast"],
+    zone: "Europe/Samara",
+    flag: "RU"
+  },
+  {
+    country: "Russia",
+    subdiv: ["Bashkortostan", "Perm Region", "Chelyabinsk Oblast", "Kurgan Oblast", "Orenburg Oblast", "Sverdlovsk Oblast", "Tyumen Oblast", "Khanty-Mansi Autonomous Okrug", "Yamalo-Nenets Autonomous Okrug"],
+    zone: "Asia/Yekaterinburg",
+    flag: "RU"
+  },
+  {
+    country: "Russia",
+    subdiv: ["Altai Republic", "Altai Region", "Omsk Oblast"],
+    zone: "Asia/Omsk",
+    flag: "RU"
+  },
+  {
+    country: "Russia",
+    subdiv: ["Khakassia", "Tuva Republic", "Krasnoyarsk Region", "Kemerovo Oblast", "Novosibirsk Oblast", "Tomsk Oblast"],
+    zone: "Asia/Krasnoyarsk",
+    flag: "RU"
+  },
+  {
+    country: "Russia",
+    subdiv: ["Buryatia", "Irkutsk Oblast"],
+    zone: "Asia/Irkutsk",
+    flag: "RU"
+  },
+  {
+    country: "Russia",
+    subdiv: ["Sakha (western part)", "Zabaykalsky Krai", "Amur Oblast"],
+    zone: "Asia/Yakutsk",
+    flag: "RU"
+  },
+  {
+    country: "Russia",
+    subdiv: ["Sakha (central part)", "Khabarovsk Region", "Primorsky Krai", "Jewish Autonomous Oblast"],
+    zone: "Asia/Vladivostok",
+    flag: "RU"
+  },
+  {
+    country: "Russia",
+    subdiv: ["Sakha (eastern part)", "Magadan Oblast", "Sakhalin Oblast"],
+    zone: "Asia/Srednekolymsk",
+    flag: "RU"
+  },
+  {
+    country: "Russia",
+    subdiv: ["Kamchatka", "Chukotka Autonomous Okrug"],
+    zone: "Asia/Kamchatka",
+    flag: "RU"
+  }
+]
